render commits on page start, show loading icon
