+++ src/components/Commits/Commits.jsx
@@ -3,22 +3,22 @@ import React, { useState, useEffect, useMemo } from 'react';
 import { 
     Box,
     Button,
-    Center,
     Text,
     Flex,
     Spacer,
     IconButton,
+    Spinner,
+    useBreakpointValue,
 } from "@chakra-ui/react";
 import { ArrowBackIcon, ArrowForwardIcon } from '@chakra-ui/icons'
 import CalendarHeatmap from 'react-calendar-heatmap';
 import { testValues, expandedTestValues } from '../../lib/dummyHeatmapData';
-import { getIssues, getCommits } from '../../utils/github';
+import { processCommitCounts, processCommitValues, getCommits } from '../../utils/github';
 import 'react-calendar-heatmap/dist/styles.css';
 import './Commits.css';
 
-
-// simple function returning color values
-const colorValue = (value) => {
+let didInit = false; // for development, makes useEffect only run once on startup
+const colorValue = (value) => { // simple function returning color class for commits
     if (!value) {
         return 'color-empty';
     }
@@ -33,60 +33,38 @@ const colorValue = (value) => {
     }
     return 'hugecommit';
 }
-// to avoid double request on render in dev
-let didInit = false;
 
 const Commits = () => {
     const [year, setYear] = useState(2024);
     const [commits, setCommits] = useState([]);
-
-    // processes github commit information into values for the react calendar heatmap library
-    // creates dict of form {date: count} then reprocesses into format of [...{date: date, count: count}]
-    const processCommitCounts = (commits) => {
-        let counts = {};
-        console.log(commits)
-        commits.forEach(
-            (commit) => {
-                let commitDate = (commit['commit']['author']['date']).slice(0,10);
-                if (commitDate in counts) {
-                    counts[commitDate]++;
-                }
-                else {
-                    counts[commitDate] = 1;
-                }
-            }
-        )
-        console.log(counts);
-        return counts;
-    }
-    // turns the count dict into an array of individual values
-    const processCommitValues = (counts) => {
-        let vals = [];
-        Object.keys(counts).map((commitDate) => {
-            vals.push({date: commitDate, count: counts[commitDate]});
-        })
-        return vals;
-    }
+    const [handlingCommit, setHandlingCommit] = useState(false);
     const commitValues = useMemo(() => processCommitValues(processCommitCounts(commits)), [commits]);
-
 
     const handleCommits = async () => {
         try {
+            setHandlingCommit(true);
             const retrieved_commits = await getCommits();
             setCommits([...retrieved_commits]);
+            setHandlingCommit(false);
         }
         catch (error){
             console.log(error);
         }
     }
 
-    useEffect(() => {console.log(commits)}, [commits]);
-    useEffect(() => {console.log(commitValues)}, [commitValues]);
-    // useEffect(() => {console.log(displayedCommits)}, [displayedCommits]);
+    // useEffect(() => {console.log(commits)}, [commits]);
+    // useEffect(() => {console.log(commitValues)}, [commitValues]);
+    // useEffect(() => {console.log(handlingCommit)}, [handlingCommit]);
+    useEffect(() => {
+        if (!didInit) {
+            didInit = true;
+            handleCommits();
+        }
+    })
 
     return (
         <>
-            <Box w="75%" h="100%" mt="5%">
+            <Box w="75%" mt="5%">
                 <Text fontSize="4xl" fontWeight="bold" color="white">{year}</Text>
                 <CalendarHeatmap
                     startDate={new Date(`${year}-01-01`)}
@@ -98,6 +76,7 @@ const Commits = () => {
                     <IconButton icon={<ArrowBackIcon />} onClick={() => setYear(year-1)} isDisabled={year <= 2021}/>
                     <IconButton icon={<ArrowForwardIcon />} onClick={() => setYear(year+1)} isDisabled={year >= 2024} />
                     <Spacer />
+                    {handlingCommit && <Spinner color="white" mt={2} />}
                     <Button onClick={handleCommits}>Display My Commits</Button>
                 </Flex>
             </Box>
